requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-docx>=1.0.1
